Description: Code alarm processing IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I5V415 Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: youbing54<youbing3@huawei.com>

+++ frameworks/graphic_math.cpp
@@ -56,10 +56,12 @@ float Sin(float angle)
     }
 #endif
 }
+
 float Fmod(float X, float Y)
 {
-    return X - (int)(X / Y) * Y;
+    return static_cast<float>(X - static_cast<int>(X / Y) * Y);
 }
+
 float Cos(float angle)
 {
 #if ENABLE_CMATH
